requirements.txt
django
django-froala-editor
pillow
readtime
psycopg2-binary
requests
python-dotenv